refactor pipeline diff components
