+++ app/cdap/components/PipelineDiff/DiffList.tsx
@@ -1,0 +1,11 @@
+import styled from 'styled-components';
+
+const StyledDiffList = styled.div`
+  background: white;
+  border: 1px solid black;
+  height: 100%;
+  width: 400px;
+`;
+
+// TODO: implement
+export const DiffList = StyledDiffList;
